tegra: host: move stale wait checking into the kernel
The kernel now receives wait tracking data (similar to gathers and
relocs) and compares the current syncpt with the threshold value.

If it's old, it gets a kernel mapping and rewrites the method data
to use a kernel reserved syncpt that is always 0 (so trivially pops
when seen by the HW).

Patch has dependency to the user-space patches

Submitted on behalf of: Chris Johnson <cjohnson@nvidia.com>
original work by: Chris Johnson <cjohnson@nvidia.com>

Change-Id: I4d4e5d3b49cab860485c4172f87247f5b4f5ea6e

Rebase-Id: R97c9d87e07991047385b014cecab9d82ab0fb0c7