video: tegra: host: separate the different kinds of ACM timeouts
Let the delay in making sure a module is idle for some amount of
time before powering it down be module specific. This typically
occurs at the end of a INTR_ACTION_SUBMIT_COMPLETLETE.

Some modules have more work (e.g. 3D) to power down, so should
have a longer delay than a module with a lower latency (e.g. 2D).
This change adds the infrastructure a subsequent change will
assign different values.

Make use of a separate #define for the timeout used in waiting
for a system suspend to go idle.

Original-Change-Id: I4d675375e7c2ad841168449bd7f6db3d38d6d6c3
Reviewed-on: http://git-master/r/40044
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>
Tested-by: Varun Colbert <vcolbert@nvidia.com>

Rebase-Id: Rdacbc9caab0bf60064d6c27a7c919b9a759da97a