Input: atmel_mxt_ts - Make wait-after-reset period compatible with all chips
The delay before the chip can be accessed after reset varies between different
chips in maXTouch family. Waiting for 200ms and then monitoring the CHG (chip
is ready when the line is low) is guaranteed to work with all chips.

v5: Fix error in timeout_counter handling
    Increase sleep time for backup check. msleep for small values is unreliable
    Increase firmware reset timeout
v4: Add a readback of command register after backup is issued, to make sure we
    are not proceeding too fast there.
v3: Add a check for NULL read_chg() function, and add the read_chg() to platform
    files using this driver (currently only mach-goni.c)
v2: At Dmitry's suggestion, add a timeout so we are not stuck looping
    endlessly in case the CHG is not going low.

Signed-off-by: Iiro Valkonen <iiro.valkonen@atmel.com>
Signed-off-by: Nick Dyer <nick.dyer@itdev.co.uk>

Rebase-Id: Re37122e88ad1afc9a937368629246690c27225b2