ARM: tegra: Increase broadcast timer clockevent rating
The new timer architecture changes the behavior of selecting the
broadcast timer. The rating of the Tegra broadcast timer (timer0)
was below that of the dummy_timer that gets selected if no other
timer is suitable for use as the broadcast timer. This causes the
dummy_timer to be selected in lieu of timer0 as the broadcast timer.
Since the dummy_timer does not have a set_next_event function, this
caused a panic when the Tegra power management code switches to
the broadcast timer. Increasing the rating of timer0 above that of
the dummy_timer below that of the the local_timer prevents the dummy
from being preferred as the broadcast timer without interfering with
the percpu timers.

Change-Id: I0666b54f07f387170b8c480adaf2f02daffcb955
Signed-off-by: Scott Williams <scwilliams@nvidia.com>

Rebase-Id: R4a3505a6320480022b888f02c6d1611cc21b08b0