asoc: codecs: max98088: check for revision id
check for revision id, it must be 0x40 as per the datasheet,
if its not 0x40 then dont allow the card to register by failing
in the codec probe. This functionality will be required when the
same board support multiple codecs. Also register the irq only
after validating the codec.

Change-Id: I8553b85d534428c7137ec2d03e2f814b898609a6
Signed-off-by: Nikesh Oswal <noswal@nvidia.com>
Reviewed-on: http://git-master/r/90061
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Scott Peterson <speterson@nvidia.com>

Rebase-Id: Rc8dcc2b38d55173a7d863a4730679733dccea042