usb: gadget: tegra_udc: Enable DR IRQ in udc_start
Currently dr_controller_run() is called only during otg state
change.
When we unload a gadget driver, for example g_mass_storage
gadget driver, dr_controller_stop() is called but when it is
loaded back dr_controller_run() won't get called becasue of the
if condition failure. i.e
	if (!udc->transceiver) {
		dr_controller_run(udc);
	}

So when the gadget driver is loaded for the second time
without removing the cable (i.e without otg state change),
the device won't get enumerated on the host because DR
IRQs are not enabled.

This CL fixes the above issue.

Bug 1012787
Signed-off-by: Preetham Chandru <pchandru@nvidia.com>

Change-Id: Ia127630e24dbe4c1e418163c2b858efb3baa8e45
Reviewed-on: http://git-master/r/114592
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>

Rebase-Id: Ref99ba89cd0b9c59e5b59c2d7846fbc1ff5fc775