mmc: host: Add stub runtime power management
Add stub runtime_pm calls which go through the flow of enabling
and disabling but don't actually do anything with the device itself
as there's nothing useful we can do. This provides the core PM
framework with information about when the device is idle,
enabling chip wide power savings.

Bug 887327

Change-Id: I1a67038d34aef2555b17c526bcdcab8177bf4bc3
Signed-off-by: naveen kumar arepalli <naveenk@nvidia.com>
Reviewed-on: http://git-master/r/125244
Reviewed-by: Varun Colbert <vcolbert@nvidia.com>
Tested-by: Varun Colbert <vcolbert@nvidia.com>

Rebase-Id: R8fba211fa97efa547ddf78eed07792d06288be84