Vibrator : Make regulator enable & disable paired
We can see many WARNING messages, when Kai try to enter to
LP0.
Because Kai didn't defined a regulator used by vibrator device
then vibrator device try to get the regulator resource, its
will got the dummy regulator. But other devices also get the
dummy regulator resource.
When a process want to turn off the vibrator, vibrator driver
use regulator_is_enabled to check wether regulator is on, if
it is, then disable the regulator, regardless the regulator
enable and disable is paired or not. This isn't right when
vibrator not use a dedicate regulator.

Move vibrator device registration in corresponding files.
Add a flag to make regulator enable and disable paired.

bug 1030465

Change-Id: I124584eac8a3a9962ee39c51020174e09d18fc69
Signed-off-by: Hayden Du <haydend@nvidia.com>
Reviewed-on: http://git-master/r/131060
Reviewed-by: Rohan Somvanshi <rsomvanshi@nvidia.com>
Tested-by: Rohan Somvanshi <rsomvanshi@nvidia.com>

Rebase-Id: R83712fcb9b6ea9d2dffe3945f0c71b9437963024