chromeos: config: enable PM_DEVFREQ
Recent kernel changes have made it required to enable PM_DEVFREQ which
is used for scaling in nvhost. This is required to be enabled otherwise
the kernel build fails with warnings about unused functions. This only
occurs if PM_DEVFREQ isn't enabled AND T114 support (which uses devfreq)
is also not enabled.

Also, renormalized configs to tot 3.4 today which pulled in a couple
more options.

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Change-Id: I9986836c43d3bbda5ba7597c4fba64ff38f06f1c
Reviewed-on: http://git-master/r/132287
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Allen Martin <amartin@nvidia.com>

Rebase-Id: R7254099f0d05d7fe5cb9d2fee13cc65c17858daf