drivers: video: tegra: refine gk20a gr idle wait
Check fifo engine status for gr ctxsw.
gr is NOT idle if ctxsw is active OR gr engine is busy.

Caller of gr_gk20a_wait_idle need to provide a value for polling
delay. This value is doubled each time a polling failed but the
max delay is limited by a pre-defined value.

Use same delay increment for other idle pollings in gr code.

Change-Id: Ie5751381f91ea7ec2530a61de6e6c076aa0a4f0f
Signed-off-by: Jin Qian <jqian@nvidia.com>
Reviewed-on: http://git-master/r/142547
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jeff Smith <jsmith@nvidia.com>