ARM: tegra11x: Control power gating of each CPU
Currently, LP2 can be enabled or disabled for either CPU0 or CPUn,
this is good for T30, but for T114, CPU0 is no different from any
other CPUs in terms of hardware power management support. Add
finer control so that power gating can be enabled or disabled for
any individual CPU in fast cluster.

The control is done through sysfs interface. The node
"cpu_power_gating_in_idle" is a CPU mask. the 5 LSB represent 5
CPUs. bit 0 is for CPU0, bit 1 is for CPU1, bit 2 is for CPU2,
bit 3 is for CPU3. By default the mask is set for all CPUs. Writing
a '1' to bit n enables power gating of CPUn. any other bits are not
used. The special case is bit 4, which controls the single CPU in
slow cluster.

For example, to enable power gating for all CPUs in fast cluster,
just write a "15" to that sysfs node. To disable power gating for
all CPUs, write a "0". To enable LP2 for slow CPU, write 0x10 to
that node.

Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/144136
(cherry picked from commit bbfd7d15eba22492d56831e4db1181128ffa56ae)

Change-Id: I12299718d286f51d8340c0258cbd1265f3212655
Signed-off-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-on: http://git-master/r/146485
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bo Yan <byan@nvidia.com>
Tested-by: Bo Yan <byan@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>
GVS: Gerrit_Virtual_Submit

Rebase-Id: Ra21713c217770922349b19cf5a27fcd41b5a8dcc