Add nvdumper kernel module to monitor reboots
The nvdumper kernel module tracks whether the system has been rebooted
cleanly.  It does this by writing 'dirty' to a fixed physical memory
address when the kernel starts.  Then, on a planned reboot, we write
'clean' to this location.  The bootloader can then examine this location
and see if the reboot was dirty or clean.  It will dump the contents of
memory after a dirty reboot.

Bug 895895

Change-Id: I31e552da9bf6218cefb77f76f7488fcf9a9151b1
Signed-off-by: Joshua Cha <joshuac@nvidia.com>
Reviewed-on: http://git-master/r/161969
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>