ARM: tegra11x: Fix race condition in cpu hot plug
The secondary CPU power up can happen so fast that it enters power
gating even before the CPU that's trying to bring it up gets a
chance to check status in PMC register. This can happen because
interrupt and preemption are not disabled, so the primary CPU can
switch to do something else, when it comes back to check PMC status,
it's already too late. When this happens, it will simply declare a
timeout failure, though the secondary power up is actually successful.

The fix is to check online status as well, because the CPU that's
coming up will set online flag before entering idle loop.

Another enhancement is to skip the status check altogether if this
is not the first boot up after cold boot or warm boot. The
synchronisation done in __cpu_up is more adequate.

For the first boot after cold boot or warm boot, set the timeout
back to 100ms.

Bug 1169512

Change-Id: Ie4b873d853c40207c7b1bc1acf0982a7bf0ecd9f
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/164065
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
Tested-by: Peter Boonstoppel <pboonstoppel@nvidia.com>
Tested-by: Sang-Hun Lee <sanlee@nvidia.com>