ARM: tegra11x: Simplify cpu hotplug sequence
There is no need to clear CPU CSR register. The "ENABLE" bit (bit 0)
is set by software, but cleared by hardware once the sequence is
completed. In case of failure, it's desirable not to change it for
debug purpose.

After CPU_CSR clear is removed, the power mask is set by the CPU
that's booting up.

Change-Id: I4879c64ec5858ed85008fbeeedb8232e35643a8c
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/164891
Reviewed-by: Rohan Somvanshi <rsomvanshi@nvidia.com>
Tested-by: Rohan Somvanshi <rsomvanshi@nvidia.com>