arm: tegra: soctherm: Fix section mismatch warnning
The soctherm_fuse_read_tsensor function was marked with __init macro,
but it's referenced by soctherm_init_platform_data function which is not
initialization function.
This is violation of __init macro usage in kernel. The __init macro
should be marked on initialization function, because functions __init
marked will be located in special section(.init.text) and this section will
be freed after boot is complete.

Bug 1176013

Change-Id: I5d042040f585b71b112d546226fd3c185c52e427
Signed-off-by: Jinyoung Park <jinyoungp@nvidia.com>
Reviewed-on: http://git-master/r/164970
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>