ARM: tegra: fuse: fusetime configuration
The fuse programming pulse needs to be 10us for T30, and 11us for T114
and later chips. Thus FUSE_FUSETIME_PGM2_0 should be modified based on
the OSC frequency used on the specific board.

The formula is FUSE_FUSETIME_PGM2_0 = N * OSC in MHz where N = 10 for T30
and 11 for T114.

Bug 1157054

Change-Id: I725eb750db19bf4253f3a77fb9518f73b4a3d9c9
Signed-off-by: Ken Chang <kenc@nvidia.com>
Reviewed-on: http://git-master/r/165303
Reviewed-by: Mrutyunjay Sawant <msawant@nvidia.com>
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>