input: misc: Invensense: fix warnings from gcc 4.7
Gcc 4.7 throws a few warnings. First is around the inline function
and removing the explicit inline seems to fix that. The other
warnings pertained to possbile use of uninitialized variables, that
I believe to be invalid given the logic, but none the less breaks
compliation with warnings treated as errors.

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Change-Id: I21e4e33e4294262fab969f2e36f19b5fd3e315fa
Reviewed-on: http://git-master/r/166263
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Robert Collins <rcollins@nvidia.com>
Reviewed-by: Xiaohui Tao <xtao@nvidia.com>
Reviewed-by: Erik Lilliebjerg <elilliebjerg@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Thomas Cherry <tcherry@nvidia.com>