arm: tegra: Dalmore: ldo9 warm boot state fix
Palmas regulator ldo9 is not in correct state
needed to power on sdmmc3 after a soft reboot.
As a result card is not detected after a soft
reboot. fix is to mantain the last software
set value to the ldo and not the OTP value
after a soft reset.

Bug 1180954

Change-Id: I2a6b34e0b13b57ba5b5bcba4c4d979778cc8a338
Signed-off-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-on: http://git-master/r/188634
(cherry picked from commit 717a91a526addbc9ce951cb33fb8a0e8ff089630)
Reviewed-on: http://git-master/r/189482
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>