drivers: #include sharable mach-tegra (TEMP)
To share code between arm and arm64, we have to relocate
sources from arch/arm/mach-tegra into drivers. As this is
not done in the main branch but our local branch, we employ
this "symlink" method to avoid merge nightmare.

The generation of these files are all scripted. Note that
in this first round we added only the minimum set of files
which are built w.r.t the tegra_bonaire_android_defconfig.

Bug 1237330

Change-Id: I636013a8a0542697f6d28097da6ba5c9a549d639
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/196602
Reviewed-on: http://git-master/r/211298
Reviewed-on: http://git-master/r/219556
Reviewed-on: http://git-master/r/224272
Reviewed-on: http://git-master/r/226411
Reviewed-on: http://git-master/r/250135
Reviewed-on: http://git-master/r/257174
Reviewed-on: http://git-master/r/249921
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>