arm:misc: Change tfa9887 driver debug spew
Changed the driver for the tfa8997 device so that
it only prints a message for the SetParam function
if there is an error and audio is playing. If audio
is not playing access to the device is not possible
so an error is expected.

Bug 1233775

Change-Id: I5e3b4f7958104d28c74dac730ea936efed567325
Signed-off-by: Scott Peterson <speterson@nvidia.com>
Reviewed-on: http://git-master/r/198963
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Niranjan Wartikar <nwartikar@nvidia.com>
Reviewed-by: Shaoming Feng <shaomingf@nvidia.com>
(cherry picked from commit 9bd34a41737e49321db885b48ee299976e1cb09f)
Reviewed-on: http://git-master/r/213741
GVS: Gerrit_Virtual_Submit