ARM: tegra: pluto: enable MAXFLASH
Program MAXFLASH_TH: The voltage of VDD_SYS that will trigger a MAXFLASH
system lock up prevention control sequence. Set this to 3.4V, the most
conservative setting. This value may perhaps be lowered once the behavior
of the MAXFLASH system is characterized.

Program LB_TMR_F: This is a timer that will trigger a re-evaluation of the
MAX-FLASH conditions after the FLASH current has been reduced. Set this to
256usecs.
Program LB_TMR_R: This is a timer that will trigger a re-evaluation of the
MAX-FLASH conditions after the FLASH current has been reduced. Set this to
256usecs.

Program MAXFLASH_HYS: set this to 200mV

Program the MAXFLASH_EN: Enable the system once it is programmed correctly.

bug 1239750

Change-Id: I66223cbeeed392b3959b6936e83763182622e893
Signed-off-by: Charlie Huang <chahuang@nvidia.com>
Reviewed-on: http://git-master/r/204274
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sivaram Nair <sivaramn@nvidia.com>
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>