arm64: Set the page table freeing ceiling to TASK_SIZE_64
In compat mode on 64bit platforms, the initial stack_top is at the top
of process VA space bound by TASK_SIZE(_32). The pages for this stack
are not getting free'ed when user process is torn down because the
ceiling is set to TASK_SIZE(_32) causing free_pgtables() to miss those
pages.

This change sets the reclaim ceiling to be TASK_SIZE_64 to cover the
entire address space, below and beyond TASK_SIZE(_32).

Change-Id: I9e09703b81a39e6d1e991556f3ad3985e7dfb9c8
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/217956
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>