ARM: tegra: introduce manual config table for t11x
To facilitate and keep track of software changes on pinmux table, manual
configuration table is introduced so that the developers are aware of pinmux
settings that are not in sync with the spreadsheet. Any pinmux configuration
changes made in software side should be upstreamed to spreadsheet.

Some refactorings are done to reduce the redundancy and new macros for gpio and
unused pins are introduced as well to be sync'd with the current autogeneration
tool.

Bug 1256207
Bug 1231210

Change-Id: I39e23cd335d36b8e39a1b7848df65d261ddbd24d
Signed-off-by: Andy Park <andyp@nvidia.com>
Signed-off-by: Ashwini Ghuge <aghuge@nvidia.com>
Reviewed-on: http://git-master/r/220389
Reviewed-by: Simone Willett <swillett@nvidia.com>
Tested-by: Simone Willett <swillett@nvidia.com>