ARM: tegra: emc: Add DRAM thermal monitoring code
For certain DRAM chips we can monitor the DRAM temperature by reading
MR4. When this register reports a particular threshold the refresh
rate and possibly the timing parameters must be updated to ensure
correct operation of the DRAM cell.

This patch provides a simple driver that periodically polls the MR4
register and if particular thresholds are hit will tell the EMC driver
to update the timings/refresh rate.

Right now only a 4x refresh mode is used; however, it is capable of
updating the timings with minimal code addition if necessary. It should
also be possible to lower refresh intervals with this driver if the
system detects that the temp is low and the EMC clock frequency is low.

Change-Id: I813e58bb3d1b3641894d04d8850c194f9a8754bc
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/221129
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-by: Diwakar Tundlam <dtundlam@nvidia.com>