video: tegra: host: Fix race from acm
The code did not take the client list mutex while setting a new
devfreq rate. However, the code utilised the list to determine
the required rate and therefore caused a race. This caused illegal
memory accesses if the client list was modified at the same time
the devfreq rate was set.

Change-Id: Ic4747e5c0fb38b717b5addcffd2d5ef7c188feea
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/224836
Reviewed-by: Mandar Padmawar <mpadmawar@nvidia.com>
Tested-by: Mandar Padmawar <mpadmawar@nvidia.com>