dvfs: tegra: avoid busy-wait during DFLL tune_high
When the CL-DVFS driver tunes the DFLL to the high-voltage range, it
must wait for the PMIC hardware to slew the voltage rail to the start
of the high voltage range.  It did this by busy-looping for 100
microseconds with IRQs disabled.  For interrupts handled by that CPU,
this presumably results in an additional 100 microseconds of
worst-case interrupt latency.  Also, during the busy-loop, the CPU
can't do any other useful work.

Fix this by releasing the spinlock and allowing the CPU to sleep or do
other work while it waits for the PMIC voltage to slew.

The downsides of this patch are that it might cause the high voltage
tuning process to take slightly longer than 100 microseconds, since
the CPU may be doing something else when the tune_high timer expires;
and also, some additional complexity in tune_timer_cb().

Bug 1282480
Bug 1632845

Change-Id: I06b9dfcfafe39c11195dedc605c4330090e30e8f
Signed-off-by: Paul Walmsley <pwalmsley@nvidia.com>
Cc: Aleksandr Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/600830
(cherry picked from commit 80ffac8f77a0f4431fdaff852dbf56bbb55df728)
Reviewed-on: http://git-master/r/735832
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>