arm64: mm: set non-cachable attrib for iomap
Memory attributes control whether how the memory access is cached.
For shared device (eg. UART), we need to turn on the DEVICE_nGnRE
attribute which instructs CPU to not cache such MMIO accesses.

When we move to full device tree model, we will automatically get
the correct flags because ioremap is called to setup the mapping
and ioremap applies DEVICE_nGnRE behind the scene.

Change-Id: Iba73d0760c073fced8ce12eb3fb6b88c3f7b6b21
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/228485
Reviewed-by: Alexander Van Brunt <avanbrunt@nvidia.com>