ARM: tegra: emc: Fix DSR override
A kernel panic was triggerable if the EMC tables were not loaded and
a device requested DSR to be reset to normal. In the case that the
tables do not exist - maybe got rejected onload - we can't reset to
"normal" because there is no normal.

Bug 1294214

Change-Id: I3e02de38e67d4979f462c759ffd070297b5aa239
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/233797
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>