ARM: sched_clock: Return suspended count earlier
If we're suspended and sched_clock() is called we're going to
read the hardware one more time and throw away that value and
return back the cached value we saved during the suspend
callback. This is wasteful. Let's short circuit all that and
return the cached value as early as possible if we're suspended.

Bug 1399318

Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from commit ffbfb5e316f0db486798ccf1db36a577ffe79637)
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Change-Id: I618d4d0c03b8ff5e5da47940fc5cfa66bc5145b1