usb: xhci: tegra: remove power saving change
We are seeing issues both with HS and SS devices on Ardbeg due to power
saving changes done when SS alone or HS alone is connected.

When SS alone is connetced we give back port ownership to SNPS to save
power (not needed on T124). At this time if HS is connected to second
port on Ardbeg board then it is not detected.

Similarly when HS alone is connected then we enable SS pad OVRD mode to
save power (needed in T124 ?). At this time if SS is connected to second
port on Ardbeg board then enumeration fails.

Bug 1301052

Change-Id: I49354df02c5af5dbec424027a3cd6e948f7ef810
Signed-off-by: Ajay Gupta <ajayg@nvidia.com>
Reviewed-on: http://git-master/r/243214
Reviewed-by: Ashutosh Jha <ajha@nvidia.com>