fs: limit filesystem stacking depth
Add a simple read-only counter to super_block that indicates deep this
is in the stack of filesystems.  Previously ecryptfs was the only
stackable filesystem and it explicitly disallowed multiple layers of
itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

Bug 200105889

Change-Id: I628881b6d2b947115ea64da380eadbd66ed4076d
Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Signed-off-by: Seshagiri.Holi <sholi@nvidia.com>
Signed-off-by: Vipin Kumar <vipink@nvidia.com>
Reviewed-on: http://git-master/r/743741
(cherry picked from commit 6d636c05ccc8eb0414a8b43fb2512bd2c1ce0d40)
Reviewed-on: http://git-master/r/736047
GVS: Gerrit_Virtual_Submit
Reviewed-by: Pankaj Kumar <pankumar@nvidia.com>
Reviewed-by: Sumeet Gupta <sumeetg@nvidia.com>