power: battery-charger: add API to acquire/release wakelock
This is require to avoid system suspend when charging is in
progress with high current. In this case, it is is require to
acquire wakelocks.

Add APIs on battery charger framework to acquire/release the
suspend.

Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/244697
(cherry picked from commit bd928b9098917e5d8716e291d663e34e37f5cd5b)

Change-Id: Ia24e65cfd37dc22aace9d41a3ad593816155645a
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/253368
Reviewed-by: Automatic_Commit_Validation_User