ARM: tegra: dvfs: Use fixed voltage as nominal.
On entry to suspend or when disabling rail scaling, used fixed voltage
instead of nominal for fixed regulator. It allows to avoid error on
setting nominal voltage above fixed level, and it is safe, since the
pll mode frequency is already capped to match fixed voltage level, and
in dfll mode voltage is always scaled to match the frequency.

Change-Id: I8e55f1c1bb551df76b2a3901fe1020857c4301e6
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/246901
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mitch Luban <mluban@nvidia.com>