video: tegra: Fix cmdbuf recycling
The lack of curly brace caused the return statement to always be
executed since the dbg_info macro always expands to a statement.

In addition, the code that was dead was broken: it passed an arbitrary
list node into a list iteration macro, which caused the list head, which
is not an entry of type struct priv_cmd_entry *, to be treated as such
as entry.  This change fixes the code to use the standard
list_for_each_entry and list_for_each_entry_safe_continue macros instead
of attempting to open-code the logic.

Signed-off-by: Robert Morell <rmorell@nvidia.com>
Change-Id: I2cffdc5a0989dc01a1acc197af8c0df6d5ef5578
Reviewed-on: http://git-master/r/250046
Reviewed-by: Ken Adams <kadams@nvidia.com>
Reviewed-by: Kirill Artamonov <kartamonov@nvidia.com>
(cherry picked from commit 4ff00d0189879b9b895fea8bb4c3d9b1260c5811)
Reviewed-on: http://git-master/r/250457
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit