sched_clock: Fix integer overflow
The expression '(1 << 32)' happens to evaluate as 0 on ARM, but
it evaluates as 1 on xtensa and x86_64. This zeros sched_clock_mask,
and breaks sched_clock().

Set the type of 1 to 'unsigned long long' to get the value we need.

Bug 1399318

Reported-by: Max Filippov <jcmvbkbc@gmail.com>
Tested-by: Max Filippov <jcmvbkbc@gmail.com>
Acked-by: Russell King <rmk+kernel@arm.linux.org.uk>
Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from commit 53c035204253efe373d9ff166fae6147e8c693b6)
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Change-Id: I9bbe733430f9f7850e4073feef4265f626e6dc8f