sched_clock: Use seqcount instead of rolling our own
We're going to increase the cyc value to 64 bits in the near
future. Doing that is going to break the custom seqcount
implementation in the sched_clock code because 64 bit numbers
aren't guaranteed to be atomic. Replace the cyc_copy with a
seqcount to avoid this problem.

Bug 1399318

Cc: Russell King <linux@arm.linux.org.uk>
Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>
(cherry picked from commit 85c3d2dd15be4d577a37ffb8bbbd019fc8e3280a)
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Change-Id: Iff7ec4aca829452c5b51576dae9437a37ed8f619