ARM: tegra: cpuidle: Attach MC clock PM on CPU0
Attach MC clock PM with CPUIDLE driver of CPU0.When attaching the
PM with cpuidle through the API pm_genpd_attach_cpuidle,it attaches to
the curent cpu's cpuidle driver on which the code is executing.Hence
there is possiblity that it can attach to other CPUs than 0.On Tegra,
CPU0 can't  enter an idle state until all the other CPUs are idle.
Therefore, the code can be scheduled to run on CPU 0.

Bug 1331433

Change-Id: I8850fb88f328723e4a286693cfc978ca8c44ca51
Signed-off-by: Ajay Nandakumar <anandakumarm@nvidia.com>
Reviewed-on: http://git-master/r/253275
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Dan Willemsen <dwillemsen@nvidia.com>