ARM: tegra12: dvfs: Add GPU cvb dvfs
Added continues virtual binning (cvb) for Tegra12 GPU. The cvb dvfs
table specifies a predetermined set of frequencies, and polynomial
coefficients to calculate minimum core voltage required for GPU to
run at the respective frequency on the part with particular speedo
value.

During dvfs initialization the cvb dvfs table is converted to legacy
dvfs format (that specifies maximum frequencies for a predetermined
set of voltages). Hence, no changes in run-time dvfs operations are
needed.

Actual cvb coefficient for Tegra12 are not available, yet. So, in this
commit cvb tables are artificially constructed to match existing legacy
dvfs limits for GPU.

Change-Id: Ied687f32cbd372b10f04486ad122ca62ba0f6b3a
Signed-off-by: Alex Frid <afrid@nvidia.com>
Reviewed-on: http://git-master/r/263701
Reviewed-by: Prashant Malani <pmalani@nvidia.com>
Tested-by: Prashant Malani <pmalani@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>