arm64: kernel: add structure to define cpu context layout
In order to define registers that have to be saved and restored on
cpu suspend, this patch adds to the kernel a structure declaration
that defines registers that needs saving and restoring across power
cycles.
asm offsets for size and members of the struct are generated so that
assembly code can make use of them in the cpu_{suspend}/{resume} code
path.

Change-Id: I9dbd52b7fb5c97a04d5a16261a8a9c1fa861d5d4
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/275931
Reviewed-by: Bo Yan <byan@nvidia.com>