arm64: kernel: implement debug monitors CPU PM notifiers
When a CPU is shutdown either through CPU idle or suspend to RAM, the
content of debug monitor registers must be reset or restored to proper
values when CPU resume from low power states. This patch implements a
CPU PM notifier that allows to restore the content of debug monitor
registers to allow proper suspend/resume operations.

Change-Id: Ibd2ee14694ea268d0d6d7b12aaa83971a2cb96fe
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/275935
Reviewed-by: Bo Yan <byan@nvidia.com>