arm64: kernel: implement HW breakpoints CPU PM notifier
When a CPU is shutdown either through CPU idle or suspend to RAM, the
content of HW breakpoint registers must be reset or restored to proper
values when CPU resume from low power states. This patch adds debug register
restore operations to the HW breakpoint control function and implements a
CPU PM notifier that allows to restore the content of HW breakpoint registers
to allow proper suspend/resume operations.

Change-Id: I9961b7b6362c4801263d0cdafbf6e63ed8cc84a0
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/275937
Reviewed-by: Bo Yan <byan@nvidia.com>