arm64: kernel: add cpu_{suspend}/{resume} build infrastructure
To enable the suspend infrastructure for the ARM64 kernel the
corresponding Kconfig and Makefile entries must be updated.

This patch adds a menu entry for power management options and
entries to enable newly added suspend/resume implementation.

Makefile changes add appropriate entries to include corresponding
files in the kernel image.

Change-Id: Ica6d2d1b9f12ceb942defa7d69f349d39603ef5d
Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Peng Du <pdu@nvidia.com>
Reviewed-on: http://git-master/r/275938
Reviewed-by: Bo Yan <byan@nvidia.com>