video: tegra: hdmi: Write complete eld buffer
The client for the eld buffer in hda_eld.c expects a
completely valid eld buffer.  The old code was only writing
certain entries and leaving other entries in the eld buffer
uninitialized.  When snd_hdmi_get_eld() then tries to read
all the bytes, it would find an invalid entry, error out,
and cause a retry (up to 8 it appears, which of course will
just keep failing).

Now we write all the entries in the buffer, so no errors
in the log like:

<6>[  192.361601] HDMI: invalid ELD data byte 31
<6>[  192.368572] HDMI: invalid ELD data byte 34
<6>[  192.378605] HDMI: invalid ELD data byte 1

and no more retries.

Change-Id: I667eb19b4d3782b4bf7805d5b095bea05214dda3
Signed-off-by: Mike J. Chen <mjchen@google.com>
Signed-off-by: Xia Yang <xiay@nvidia.com>
Reviewed-on: http://git-master/r/354010
Reviewed-by: Automatic_Commit_Validation_User
Tested-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>