video: tegra: host: gk20a: Abort failing engine
The interrupt handler used to read fecs to get the current channel
and assumed that it was causing the failure. However, usually the
failure occurs when we are trying to save out context of a busy
engine and therefore the failure is not in the current context.

This patch modifies sched error interrupt handler to find the engine
that is failing. In addition, the interrupt is cleared to allow
new sched errors occur.

Bug 1351268

Change-Id: Icc1714f8b03e404994553e49d9830afcf41172ac
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/272021
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>