extcon: palmas: confirm cable state with line state
It is observed that USB sub module of the Palmas device does not
show the correct cable state of USB-ID and but line state of
the ID line shows correct state.

Hence before updating the cable state, reconfirm it based on line
state register.

bug 1367328
bug 1367205

Change-Id: I16975eaf76038c3a2b9f378b0e1e1081f01eb229
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/273836
GVS: Gerrit_Virtual_Submit