video: tegra: host: gk20a: No PMU on MMU fault
ELPG must be disabled during engine reset to make sure that the engines
are not turned off (holding the mutex is not sufficient). However, simply
enabling ELPG afterwards crashes both PMU and GR and therefore better
approach is to simply disable PMU in the beginning of the MMU fault ISR
and turn on the PMU back at the end of the ISR.

This patch makes the MMU fault code to disable PMU and enable
it afterwards. To make sure that PMU initialisation works nicely,
the code keeps runlists empty until the PMU is powered back on. Otherwise
PMU may fail to boot. This behaviour is similar as what the code does in
resume handler.

In addition, PMU initialisation sequence now waits for PMU idle before
proceeding.

Bug 1342089
Bug 1365804
Bug 1361532

Change-Id: I43d67539dcc6b28fba6099571e244ac103f8a8ff
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/276162
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>