ARM: tegra: ardbeg: powermon: set ina230 avg value
Sets a value for the avg field of the ina230 config.  This will take 128
samples per-reading.  With the default conversion time of 140us, this
works out to ~18ms per-reading, which is a long time, but works well for
anti-aliasing power measurements for changes in power consumption over
graphics frames (~16ms).

Change-Id: I81f2a40dbae3f70e2fb24538d5a01b5c7c17ac24
Signed-off-by: Jon McCaffrey <jmccaffrey@nvidia.com>
Reviewed-on: http://git-master/r/299577
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>