video: tegra: host: add slave_initialized flag
A slave platform_device will be registered when we initialize the
master device. But if modprobe to init master device again, it will
try to initialize slave platform device again then oops happens.

And if we registered a platform_device in kernel, we normally don't
release it. So introduce a flag of nvhost_device_data to indicate
the status of slave and skip the slave init.

Bug 1377330

Change-Id: Ia4dfe6f8046952e2642343d23e39a14b6eb021b4
Signed-off-by: Bryan Wu <pengw@nvidia.com>
Reviewed-on: http://git-master/r/289326
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>